test: split pmemobj_log* tests to separate folders
