+++ src/test/obj_log_get_treshold/.gitignore
@@ -1,0 +1,1 @@
+obj_log_get_treshold
