+++ src/test/obj_log_get_treshold/TESTS.py
@@ -16,12 +16,12 @@ class OBJ_LOG(t.BaseTest):
     test_type = t.Short
 
     def run(self, ctx):
-        ctx.exec('obj_log_set_function', self.test_case)
+        ctx.exec('obj_log_get_treshold', self.test_case)
 
 
 class TEST0(OBJ_LOG):
-    test_case = 'test_log_set_function'
+    test_case = 'test_log_get_treshold'
 
 
 class TEST1(OBJ_LOG):
-    test_case = 'test_log_set_function_EAGAIN'
+    test_case = 'test_log_get_treshold_EAGAIN'
